fea(utils/encap2) new version of encap
This new version of encap is a partial rewrite of the old version. Code has been copied from the old version.

The reason for the rewrite is because the old version has a complex logic (hard to maintain) and encap has introduced new features to the protocol.
The new version is compatible with the persistent storage of the old version.

Usage:
Use EncapClient:: registerDevice(), authenticateDevice([optional authenticationId]) and saveDevice() to authenticate to encap.