fix(agent/utils): Made function handle key pair abstractions
This fixes a bug in the Digipass crypto module.
CryptoUtils::calculateSharedSecret would take a
Bouncycastle ECPrivateKey as static type rather
than its abstraction, java.security.PrivateKey.
Since the Digipass module would downcast
ellipticCurveKeyPair.getPrivateKey() to a
ECPrivateKey instance in order to make a call to
this function, it was required for
ellipticCurveKeyPair to be a Bouncycastle object
at run-time. This requirement would be violated
when we deserialized a string into a
sun.security.ec.ECPrivateKeyImpl instance.