feat(core/account): Making setAlias mandatory
API-wise, alias is a mandatory field since it
is always set to something no matter what. The
value is either set using the optional setAlias,
or falls back to accountNumber if it is not set
(or if it is set to null). This makes sense
because alias is literally what is displayed to
the end user.

However, the builders do not reflect this, since
setAlias is optional. It is our intention to force
the implementor to decide what the alias should
be, rather than relying on implicit behavior.