fix(Swedbank): Search for transfer src account in all profiles
Previously we've only considered the latest profile we've seen. This causes issues for users with multiple profiles who choose to transfer from an account that belongs to a profile we didn't select (we say the source account can't be found then). This method will loop through all profiles to find the source account of the transfer, and then select that profile to be used going forward in the execution flow.