feat(SE SEBKort): migration to sanitise card numbers for SEBKort agent
This commit adds a migration to sanitise card numbers to match the
format required for the nxgen SEBKort agents. This commit will only add
the migration, with a `changeRequest` implementation that will cause
the new agent to be used if data has already been migrated, but no data
will be migrated until the new provider config has been deployed.