feat(nasa): Refresh of aggregation from NASA (only used in test)
We now reach aggregation through apache http client and trigger a refresh.
This is only used in unit test for now expecting aggregation to be running on a local minikube environment.
The solution needs refinement, but makes sense to do that after we can create credentials to be able to trigger a real fetch of account.