fix(SEB): Set correct name of fund account
We should use the account type, as the name field is a fund name. Using fund name has caused confusion as the investment account will get the same name as one of its instruments. I verified with one user that the accountType is "Fondkonto" which is a much better choice for account name.