feat(danske): authenticate using code app if it's the preferred device
This commit contains refactorings to the existing challenge authenticator to support both keycard (previous) and code app. The code app authentication will be initiated if the preferred device is code app (which it is per default if the user has NemId installed on a device).
Once the code app auth is initiated a push notice will be sent to the user's device.