feat(payment-revamp): Adds PaymentType, skeleton for exceptions
Payment type [SEPA, DOMESTIC, INTERNATIONAL] added to the Payment object.
Start of exception handling for Nordea although a separate PR will be created
to handle exceptions.
Ability to handle several PaymentExecutors, now there will be a payment controller
created for each request to the Agent instead of the agent keeping an instance.
This allows for different executors to be created depending on the PaymentType and
it is necessary due to the fact that there might be different objects needed depending
on the payment type and/or the requests need to be sent to different endpoints
in the bank side.