refactor(einvoices): Discard einvoice dependency in unrelated agents
For the longest time, every single nxgen agent has
been forced to inherit a method named
constructEInvoiceRefreshController.
That is a violation of the Interface segregation
principle. In reality, only 4 out of the 92 agents
implement fetching of einvoices. Consequently, 88
agents are polluted with complexity they have
nothing to do with.

This change will make it necessary for agents
wishing to implement support for fetching
einvoices to implement the einvoice executor
interface, which much better from a design point
of view.