fix(danskebank): dont crash on missing owner name
DanskeBankV2Agent threw exception if an entity fetched during
authentication was not populated as expected. It seems like password
authentication sometimes return an entity that is missing LoginInfo.

This fix allows the account holder name to be empty and introduces more
granular exceptions/logging while a more permanent solution is developed.