fix(SE/DanskebankV2): Don't treat no user response as error
If the status is 0 and the status text is null it should be an indicator that the response from Danskebank is fine, so the user had every possiblitiy to input the challenge code. So if the challenge response exchange outcome is null we'll treat it as a supplementalInfoError. I.e. an authentication exception.