refactor(DK/SdcSmsOtpAuthenticator): Make error testable
Refactored the code to make part of it testable.
To be able to test the exceptions received one would have to mock a bunch of methods and still it would probably not work.
By refactoring the catch-block and making a function of it makes this part of the code testable by supplying the correct error message.