fix(skandiabanken): instrument price NPE
Some instruments does not have an `askPrice` which caused NPE.

In these cases we use the `todaysRate` of the `HoldingEntity`. This is
based on comparing how prices are presented at `morningstar.se` and
how we do it in other agents. But a ticket will also be created to
investigate this further.