feat(Sabadell/LoanDetailsResponse): Add LoanDetailsResponse
It's possible for an account to be registered for multiple loans.
However, I felt it was best to only parse one of the loans since our
codd does not allow for multiple loans to one account right now.

I added a print statement to warn when multiple loans where found, but
the code will only parse 1 of the loans.