fix(TinkHttpClient/Logging): Add logging for exception
When writing the Sabadell loan details implementation I encountered a
problem where the HTTP POST body would be `null`, even though an object
was supposed to be embedded.

The reason was that I forgot to add a @JsonObject annotation, but since
the POST failed silently and didn't send any data it took 45 minutes of
debugging to find that the exception was not logged. This PR intends to
fix this issue for other developers.