feat(agent-dat): test for tracking accounts from AgentWorkerContext
Add `CapabilityTransmitterWrapper` to `AgentWorkerContext` such that it
will send refreshed accounts to the agent-dat service.

In order to be able to test this carefully it is only configured to run
if a specific test provider (`tracking-test-provider`) is used to
trigger the refresh. This provider will only be available on
oxford-staging.