fix(refresh): mark loan account and transaction refresh deprecated
Since the addition of implementing refresh executors in the controllers we should no longer use the methods implemented from AccountRefresher and TransactionRefresher.

They are still there due to tests using those methods and expect the class to implement Refresher interface. Future refactorings have to be made to fix the tests and remove this dependency.