feat: Gitignoring etc/development.yml
Currently, you are expected to modify etc/development.yml if you want to
supply configuration values that are to be used in tests. However, since
this file is tracked by git, the git repository will be in a constantly
dirty state during development. Not only is it a pain to go through this
patch every time I'm doing `git add -p`, but there is also the risk of
accidentially committing and pushing these changes.

To work around this, I have renamed development.yml to
development.template.yml and then git-ignored development.yml. Now
whenever someone wishes to do development, they are expected to rename
development.template.yml to development.yml and then update the latter
with the sensitive values.