fix(se-handelsbanken): Corrected first available payment date
Currently the Handelsbanken agent is assuming that any day that is not a
weekend day or a Swedish public holiday is a day eligible for payment.
New Year's Eve is not a public holiday but the bank is closed on that
day, so any attempt by the agent to make a payment on that date would
fail. To solve this, the agent will from now on set the date to the
firstPayDate field which we receive earlier.