feat(agents): Allow alternative Agent constructor with config
Agents are currently always constructed using a constructor with
parameter list (CredentialsRequest, AgentContext, SignatureKeyPair).
After the agent is instantiated, the agent factory invokes
Agent::setConfiguration(AgentsServiceConfiguration). The agent may
override this method to load configuration values it needs.

In practice, this design is frustrating to work with. Many fields in the
Agent class can be dependent on configuration values, which means that
they cannot be fully initialized in the constructor. Instead, they have
to be initialized in the setConfiguration method or later. And of
course, fields dependent on those fields also have to be initialized
there, at least if they want to stay immutable. setConfiguration often
takes the role of an awkward necessary second initialization step.
Fields cannot be final and the design encourages mutability.

The proposal is to give the agent the option of using an alternative
constructor with an AgentsServiceConfiguration parameter instead of
SignatureKeyPair. Everything can then be initialized in the constructor,
and there is no longer any reason to override setConfiguration.