feat(jwtstate): add method to generate pseudo id
Each user of the JwtStateUtils generate their own pseudo id using randomutils. This change moves that logic into its own method and also uses the appUriId (generated in main) if it's set.
This change will make it possible for us to stop using the JWT in the near future (by using the appUriId as state and pseudoId).