feat(eidas): Adds Client Config template builder
The way this works is it looks into the package of the agent
class to try to find a class that implements ClientConfiguration.

If it doens't find it it looks into the package of the agent's class
superclass and so on as long as the package it is looking into is
a subpackage of the se.tink.backend.aggregation.agents package.

After that it builds the template based on the new @Secret and
@SensitiveSecret annotations and the optional getDescription...
and getExample... static methods.