feat(strongAuth): Standardize strong authentication state
It is important that the state used in strong authentication (oauth/openid) is generated the same way for all agents and controllers.
This is because the state is used to lookup information needed to finalize the user redirect in our callback (main). Thus we use the appUriId sent to us from main (generated randomly in main) as a preferred state.