fix(nl-volksbank): Moving client ID+secret to Authorization header
Client ID and secret are currently passed in the /token request as
the query parameters client_id and client_secret. However, Volksbank's
documentation states that client ID and client secret are supposed to be
included in the Authorization header. It works either way, but we'd
better comply with the docs. May be the cause of some errors we are
sometimes getting when refreshing the access token, who knows.