refactor(nxgen/agent): Pulled down authenticator to NxgenAgent class
It is not appropriate for `SubsequentGenerationAgent` to implement
`ProgressiveAuthAgent`, because only a few SubsequentGenerationAgent
agents actually use `ProgressiveAuthAgent::login`. As such, that
approach relies on runtime casts of `Authenticator` to
`ProgressiveAuthenticator` which we should avoid. The IOC flow of
execution is also quite hard to follow and freely adjust.

The solution is to move the old `Authenticator` to
`NextGenerationAgent`, then let every progressive agent implement the
`ProgressiveAuthAgent` directly. We have already done this with
refreshable item executors, and this will take us one step closer to
killing off these parent classes.