fix(DebugWorkerCommand): Add ability to configure debug log frequency
This can then be used to slowly increment the amount of debug logs generated for refreshes and transfers and monitor the system load and performance degradation when it is increased incrementally to see how far we can push it.
We need more debug logs for trouble shooting, and not just when things go really wrong.