fix(Swedbank): Adding retry for fetch methods
Swedbank generally throws 500 errors when they are not able to handle the request but when the request is retried; most of the time; it gives proper 200 response. So, we have handled the 500/504 errors for fetch methods and added a retry for 2 times more before consuming the exception. In all other cases, exception is thrown as it should be.