build(aggregation/lib): Cut off compile-time dep on agents pkg
Code related to specific agents does not belong in the aggregation/lib
package, so there should exist no such compile-time dependencies. On the
other hand, the aggregation/lib package has a necessary run-time
dependency on all agents, because it contains AgentClassFactory which is
used for creating an agent class instance from a string or a provider.
I am making this more clear by converting the inadvertent compile-time
dependency on :agents to a run-time dependency.