build(aggregation/lib): Moving up run-time agent deps to aggregation/lib
Every agent is packaged in one of two ways: either it is contained in a
single isolated Bazel package (e.g. "erstebank"), or it is part of a
common large Bazel package named "agents". The former approach is
preferable, most notably for the dramatic build time improvements.

Agents generally don't have dependencies on each other, and definitely
shouldn't have. Yet, currently the agents package has a run-time
dependency on all the isolated agent packages. This makes no sense and
implies for example that it is possible to construct an Erstebank agent
instance inside Aib agent code.

The Bazel package that _should_ have a run-time dependency on these
packages, including the agents package, is the aggregation/lib package.
This is where AgentClassFactory resides, which is responsible for
creating agent instances. This dependency used to be transitive via the
agents package, but it is now made direct.