fix(ES/ING): Temp fix date parsing for loans
As we currently don't get any logs we can't investigate the issue. I don't want the toJavaLangDate method to have the default behaviour of returning null, as it's not how it should work. Therefore added a separate method in the loan fetcher, once we get debug logs again we can fix this properly.