refactor: MultiFactorAuthenticator -> TypedAuthenticator
MultiFactorAuthenticator does nothing but extend TypedAuthenticator, so
it is a pointless piece of indirection which increases the cognitive
load in comprehending how the authenticators are implemented. If we
really want to make a point that a multi-factor authenticator is a type
of typed authenticator, it should be done via a doc comment.