fix(avanza): Parse loan details correctly.
Parsin is done from these end user instructions:
The product Superbolånet is somewhat special.
remainingLoan: 170700.14 - As said before, this is the credit line available.
totalBalanceDue: 5378947.00 - This is the current loan amount.
nextPaymentPrognosis: 3528.60 - This is the interest for the last month.
I don't amortize and I'm not aware of any possibility to set up a monthly amortization for Superbolånet. Amortization, for this product, is done ny one off payments, i e the 170k (actually two separate payments in my case).
So, if you have done the changes I pointed out in my previous response and in addition to this simply empty the amortizations field for Superbolånet specifically it will be correct.