feat(LogMasker): Cleanup and new CreateLogMaskerCommand
Cleans up a few things and introduces CreateLogMaskerCommand which
will crete a LogMasker per request.

I'd still like to improve this with DI but I think we should try to
re-enable S3 logs first.

If this works when I test it locally I'll start building the RE agents
part so their secrets are also masked.