fix(metric): don't track temp error for disabled provider
It is possible to fore `refresh-manual` for a credential that belongs
to a disabled provider. Currently this will put the credential
in `TEMPORARY_ERROR` instead.

This commit stops tracking amounts of `TEMPORARY_ERROR` for providers
that are set to `TEMPORARY_DISABLED` or `DISABLED` since these cases
will spam a lot of non-actionable alerts until the provider is enabled
again.