build: Introducing rules_jvm_external, take 2
In my previous attempt to introduce maven_install, the aggregation
service threw an exception in staging during scanning of the
aggregation_service.jar:

java.io.IOException: invalid constant type: 19 at 4

This time, I am keeping the version of the library unchanged, i.e. at
3.1.5 instead of 3.2.0. I have tested locally with minkube and was able
to reproduce the error with 3.2.0 but not with 3.1.5. I also tried
updating the version without migrating to maven_install, and got the
same exception.