chore(GPI): Handle bankId exception in payment controller
This is so that we don't have to have this exact parsing in every agent that uses bankID for signing payments. We'll probably need to rethink this logic when we have more signing cases implemented, for thirdparty apps etc. Settling for this now so we don't over engineer for a use case we don't fully understand.