fix(nordea): Parse subloan interest rate properly.
We cannot parse each sub loan as its own loan as we do not have a unique identifier per subloan, so we have to parse it as one loan until we expand our model to reflect multi-part loans. This fix will try to get the interest rate that is on each subloan, if the base rate is not null and they are all the same rate, if not throw an exception, as it is a state we do not handle, and will have to investigate further if it would arise.