fix(Nordea): Handle error when fetching payment details
Catching this issue in the api client as it's used both for payments and transfers. Also, since it's a generic error from Nordea we also want to know that it's this call that fails. If we handle this error in the larger try catch block in NordeaBankTransferExecutor.executeTransfer there may be other calls that fail with this generic description, which would make the exception info incorrect.