build: Added missing resource section to vendored aws_java_sdk_core
We currently have two conflicting versions of the same
`com_amazonaws_aws_java_sdk_core` jar in our classpath. One is a regular
old `maven_jar` declared in WORKSPACE. The other one is a vendored,
patched copy residing below `third_party/aws_java_sdk_core/`.

The vendored version takes precedence in the class path, so in theory,
the non-vendored jar is useless to us. We'd better off removing it
to avoid confusion.

However, it turns out that at runtime, at least one file from the
non-vendored jar is actually used, namely the resource
`awssdk_config_default.json`. This occurs because this file is not
present in the vendored jar. This in turn occurs because we have not
included it in a `resources` section in our vendored Bazel target.