fix(Swedbank): Don't treat 500 during tx fetching as bankservice error
For some users Swedbank don't return more than 3 months of transactions. Throwing a bankservice exception means that the credentials for those users will never be refreshed. The transaction fetcher will treat the 500 response as "no more transactions", and we'll return whatever we got. The credentials we get pagination errors for are marked so that we can backfill in the future (if Swedbank ever solves their pagination issues...)