build(se-swedbank): Moved "service provider" into agent directory
This so-called service provider does not meet the criteria of being one.
It imports a lot of code from the SE agent, so there is a nasty cyclic
dependency between the two. At first I tried to break the cyclic
dependency somewhere, but the coupling was too tight.

In addition, the SE agent is the only one using this service provider.
So the existence of such a separation is not close to being justifiable.

For now, I will move the service provider directory inside the agent. At
least that will allow me to create a Bazel package for the Swedbank
agent. The owners of this agent should consider whether it is wise to
move the classes closer to the agent classes.