refactor(storage): Moving static method out of interface
I object to the practice of cluttering interfaces with implementation
details. We are including two conceptually different things in one file:

* The declaration of a contract
* A code segment shared by two particular implementations of the contract

Only the former belongs in interfaces, unless there is a strong reason
to do otherwise. This is because the true purpose of interfaces is to
achieve abstraction, and the practice of including implementation
details inside an inteface damages the effectiveness of the abstraction.
Effectiveness is achieved through parsimony. It is also not a scalable
approach, because implementation details tend to bring in additional
dependencies on libraries, and clients of the interface must bring in
those dependencies regardless of whether they end up using the method or
not. That violates the interface segregation principle.

The better alternative is to move the code to a common class (util class
will do for now).