feat(rpc/account): support nullable balance
rpc.Account can now have a null balance. This was previously always
interpreted as zero, which makes it harder to know if a zero balance
is actually because of an empty account or a bug.

This change will enable us to track null values using the
Data Availability Tracker.

The object will still serialize balance as a primitive, interpreting
null as zero. This is a precaution since system may rely on this
behavior to some degree.