fix(revolut): allow voice mail verification
Fall back on voice mail verification if SMS channel is unavailable.

**Why?**
Revolut stopped sending SMS activation codes when first registering a
new device in favor for an email that deep-links directly into their
app. This method is not usable to us.

If we receive a deep-link email we will now attempt to resend the code
in a voice call instead.