fix(background_refresh): lower default provider rate limit
Due to the way the background refreshes run, we inadvertantly cause us
to consume from the background refresh at the same pace as we produce to
it. While this is generally fine for a system that wants to maximize
throughput, the Aggregation service also handles other things.

Specifically, during the background refreshes, the CPU of the
aggregation service gets massively throttled. This implies that the
underlying hardware of the Kubernetes Pod actually becomes resource
starved. This commit is another attempt at fixing #TPA-15, and should spread the
consumption of the background refresh over a longer period. We do this
by lowering the default rate limit by 80%.

A potential risk with this commit however is that we risk spinning on
re-adding messages to the queue, and subseqently immediately popping it
again. Code has however been found in the SqsProducer class, that on
requeue hides the message for a random amount of time (between 0-180
seconds).