refactor(agent): break agent strategy into several components
In order to be able support the many combinations of different
component providers (previously called strategies) we are currently
using the component provider is broken into several pieces that can
be selected individually. The providers with their current variations
are:

TinkHttpClientProvider:
    - NextGenTinkHttpClient
    - LegacyTinkHttpClient
    - WireMockTinkHttpClient

SupplementalInformationProvider:
    - SupplementalInformationProviderImpl (normal production impl)
    - MockSupplementalInformationProvider (to be used for mocking)

AgentContextProvider:
    - AgentContextProviderImpl (normal production impl)

A selection of these sub-components are then composited in
AgentComponentProvider which is the used to build the agent depending
on its purpose.