fix(sqs_consumer): add static maximum consumption rate
Due to the large spikes of background refreshes that happens at specific
times throughout the day, a "flattening" effect is desired to slow down
the consumption rate but keep the production rate the same. This has
been tracked in #TPA-15.

We make this change observing the following: Today we consume and produce
the entire queue simultaneously in around 16 minutes. However, the
lasting effects of the background refresh can be seen as long as 45
minutes after. This is due to the background refreshes persisting on the
pods, while waiting for the per-provider-rate-limiter to allow them
through (There are about 700-800 Runnables created during this period).

By slowing down the peak consumption rate from 5/second to 2/second,
the effect should be:
- More even load per pod.
- Lower consumption rate.
- Shorter total produduction time.