fix(background_refresh): remove async dispatch of refresh
This commit removes the dispatch from the rate-limiter. Only 10 queued
requests per provider is allowed, this means that when we're getting
anything above that, the ThreadPool will start throwing
RejectedExecutionExceptions, causing the background frefresh to be
requeued in SQS.

This change will work especially well with how System currently sends
information to us. It goes through Providers, and selects Credentials
from those. Meaning we're getting providers in batches, which would lead
to an even distribution per provider over the existing pods.