fix(amex): fix broken header
Amex does not seem to case that this header is broken, but it messed up
our tests in that the `Content-Type` header was not parsed. Having
fixed this we also needed to update the mock files to include this
header.

The change was verified using `AmexV62UkAgentTest` with ambassador
credentials, as well as both mock test implementations.