fix(nxgenhttpclient): restore single use filter
In our HttpClient interface we can add "filters" to the client itself,
and also to the `RequestBuilder`. In the legacy package these two ways
of adding filters would have the following effects:

1. Add filter to client -> Filter used for all request from that client.
2. Add filter to RequestBuilder -> Filter exists only for the request
build from that RequestBuilder.

The NextGenTinkHttpClient package shared the filter list between the
client and the RequestBuilder, resulting in all filters ending up as
persistent on the client. This is a breaking change for a few agents.

This change will pass a copy of the filter list to the RequestBuilder
instead, making behaviour compatible with legacy code.