rafactor(ukob): add new component constructor option
- Add the new `AgentComponentProvider` constructor to the
`UkOpenBankingBaseAgent` alongside the old constructors.

- Mark old constructors as deprecated.

For UK agents using the new constructors will do two things:

- Switch to the NextGenTinkHttpClient.
- Give access to the dependency injection introduced through
`AgentFactory`, which is a prerequisite for enabling automated wiremock
tests.

This commit will not affect any particular agent. In order to start
using the constructor the final implementation needs to implement it
as well.