test(aggregation/service): Added test for conflicting Guava deps
This test will analyze the classpath that results from dragging in the
aggregation service as a runtime dependency. It will then assert that
there exists exactly one version of Guava.

This will counteract cases where we have accidentally ended up with
multiple different versions of Guava on the classpath by, for example,
updating the repository's dependency on tink-backend.