build(WORKSPACE): Updated commons-collections: 3.1 -> 3.2.2
According to
```
bazel query 'deps(//src/aggregation/service)' | grep commons-collections
```
there exists two different versions of `commons-collections` on the
classpath: one from maven_jar and one from the selenium maven_install.
That's not good, because the version that will be selected at run-time
is undefined. I'm updating the version so we have a single version for
the aggregation service.