refactor(ukob/barclays): inject JwtSigner
UkOpenBankingBaseAgent now has a module that sets up dependencies
requiered for injecting JwtSigner with instance of LocalKeySigner.

UkOpenBankingBaseAgent also has a new constructor taking the JwtSigner
directly.

In order to not moving all agents to injecting JwtSigner in this commit
some duplicate code is tolerated, but this should be handled in
subsequent commits.