ci(sonar): Run neither on master nor staging
We currently allow Sonar to run on any branch that is not staging.
This means that Sonar is currently run on the master branch, which is
not desirable.

When submitting a PR, three different branches come into play:

* Your feature branch, named something like `Sebelino:disable-sonar-on-post-merges`. When you create a PR, a "pre-merge" build is triggered.
A bunch of Buildkite steps are run on this one, and they are visible on the PR in GitHub before borsing.

* The "post-merge" build, always named `staging`. This one is created by Bors and exists
to verify that the merge of your PR + the latest revision of upstream master passes all Buildkite steps.
When you `bors r+` a passing PR, Bors will trigger another build, this
time on the staging branch. If you have ever seen Bors respond with
"Build failed" after doing `bors r+`, that means the staging build
failed. All builds for the staging branch are visible here:
https://buildkite.com/tink-ab/build-tink-backend-aggregation/builds?branch=staging

* The actual trunk, always named `master`. If the staging build passes, the commits
from the staging branch will be copied over to the master branch.
This will trigger a third build, this time for master.
There is no need to run the unit tests and other checks a third time, so
that should not be done in this build. On the other hand, this build
triggers the release train to start, deploying the changes to all
environments (aggregation-staging, then aggregation-production).
All builds for master are visible here:
https://buildkite.com/tink-ab/build-tink-backend-aggregation/builds?branch=master