build: Updating to newest revision of tink-backend
Since the last time this hash was updated, a few breaking changes were
made to EventProducerServiceClient:

https://github.com/tink-ab/tink-backend/pull/19147

Specifically, EventProducerServiceClientProvider now expects an
EventProducerServiceClientChannelBuilder-annotated
ManagedChannelBuilder to be dependency injected.

This required me to add a binding to AggregationDecoupledModule, or else
the service would crash on startup due to a missing binding.

No changes need to be made to the production Guice module in this case,
because it is already using EventProducerServiceClientModule from
tink-backend, which was already adjusted for this change.