feat(update-transactions): Exclude null values from contract
As we have environments where we are not allowed to relase on a regular
bases, this change allows us to control what values are propagated
towards AC/System by setting them to null.

This is *not* the preferred way of doing this, as the deserialization is
a better place to control that. However, as delivering the necessary
changes to Cornwall will take longer time than making this change.

This can thus be seen as temporary until we've got the changes in
https://github.com/tink-ab/tink-backend/pull/19513 deployed to all
environments.