feat(decoupled): Abstracted away fake bank in decoupled mode
Instead of having the aggregation service make calls to a separate
"fake bank" process over HTTP REST in decoupled mode, a WireMock server
instance will be created on the fly every time an agent is instantiated,
then torn down when it is no longer needed.

This decreases the number of services the user needs to set up to
successfully run in decoupled mode, from 3 to 2.