build(dropwizard): Manage deps with new_git_repository instead
This is how it is done in tink-backend already. The difference between
git_repository and new_git_repository is that instead of relying on the
BUILD file in our Dropwizard fork repository, we use a BUILD file
residing in this repository. The advantage is that we can then directly
use the `aggregation` maven_install in this WORKSPACE instead of letting
this WORKSPACE depend on //jar deps found in the fork's BUILD file:

https://github.com/tink-ab/dropwizard/commit/0c2f90f4358e262d0fe0af3f6d31eb0fa3cabc40#diff-96239ee297e0a92ac6ff96a6bc434ef0R1-R28