refactor(workers): Decoupled worker operation factory from agent factory
In `AgentWorkerOperationFactory`, in the case of creating a refresh
operation, we are currently checking whether the provider found in the
`CredentialsRequest` contains a `className` field that points to a
subtype of `SubsequentGenerationAgent`. If not, we add some extra
commands to the operation. We do the check by instantiating the agent
using `AgentClassFactory`, then introspecting its type.

This is a hack, but it is hack we cannot get rid of as long as the
"prevgen" agents exist.

Increased complexity aside, it also has a severe impact on the
development feedback loop. The fact that `AgentWorkerOperationFactory`
has a dependency on the agent factory means that it needs to have all
the jars for all the agents available at load-time. This dependency
slows down execution of unit tests for this class from 4 seconds to a
crippling 14 seconds.

You can see this slowdown in action by adding a dependency on
`agentfactory`:
```
buildozer 'add deps //src/integration/lib/src/main/java/se/tink/backend/aggregation/agents/agentfactory' //src/aggregation/lib/src/test/java/se/tink/backend/aggregation/workers/worker
```
and then running a test in `AgentClassFactoryTest`.

To cut off this dependency, I replaced the static call with a
`Predicate` and dependency injected the implementation of this predicate
from the top-level Guice modules.