build: Removed empty performance hog test
I must have made a mistake with an earlier PR, because there are
no sources in this directory. Just a BUILD file with a `junit_test`. It
turns out that the mere presence of this BUILD file basically slows down
the CI pipeline by 29 seconds!

The reason is that this BUILD file has a dependency on this target:

//src/aggregation/lib/src/main/java/se/tink/backend/aggregation/workers/worker

and that target has at the time of writing a runtime dependency on ALL
agents.