refactor(Comparor): Move generic comparor from AgentContractAsserter
The aim of this commit is to move the entity comparor that we were using for making assertions on agent contract entities to outside of AgentContractEntityAssertion package. The reason is that I aim to use it for comparing request bodies of two HTTP requests as well on the next commits. The comparor is generic enough to handle comparison of both agent contract entities and HTTP request bodies. However to keep it even more generic I made the following changes after moving it out:

1) Since the definition of "difference" depends on the context / use-case the comparor class now uses a "DifferenceCounter" component which is supposed to be provided by a component that wants to use the comparor.
2) The methods that are responsible of producing an error message and the methods that are responsible of finding the actual differences are now in different classes. This new comparor package is responsible of only finding the differences, not responsible of creating error messages