feat(wiremock): test module overrides production module
Agent package modules were not loaded when running tests. If production
and test required the same module it had to be re-specified in the
test module.

Now the developer only needs to specify the compontents that should
change when running the test, meaning that you can guicefy the agent
without having to update the tests until you decide that you want or
need to inject a test dependency instead.