refactor(ukob): all agents injet JwtSigner
All UKOB agent now pass their JwtSigner to the base agent through the
constructor instead of loading it in `setConfiguration`, the main
purpose for this is to allow mocking of the `LocalKeyJwtSigner` which
is now passed via guice.

Danskebank outside of the UK still construct their signer in the agent,
but now in the constructor. This was simply done to limit the scope of
this refactor somewhat.