fix(aggregation/lib): Resurrected unrunnable tests
I noticed to my dismay that none of the tests inside this package can be
run. I don't exactly understand the details, but I think it is related
to how our junit_test rule is implemented. This problem is fixed by
moving down the BUILD file closer to the sources.