refactor(utils/mappers): Static initialization -> singleton
When investigating JDK11 compatibility, I noticed that one test would
fail because this class couldn't be found. It turned out that the reason
was that an exception was silently thrown at class initialization time
in one of the two static initializations here.

This confusion and needless computational expense led me to remake this
into a singleton class.